style: minimal style applyed
