+++ src/app/product/[id]/product.module.css
@@ -1,0 +1,7 @@
+.container {
+    display: flex;
+    width: 100%;
+    max-width: 1100px;
+    margin: 0 auto;
+    flex-direction: column;
+}
