+++ src/components/cardDetails/cardDetails.module.css
@@ -1,0 +1,79 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    justify-content: space-between;
+    gap: 20px;
+    margin-top: 30px;
+    padding: 10px;
+    height: 100vh;
+    align-items: center;
+
+    @media (min-width: 768px) {
+        flex-direction: row;
+    }
+}
+
+.info_img {
+    width: 100%;
+    height: 500px;
+    box-sizing: border-box;
+    overflow: hidden;
+
+    img {
+        width: 100%;
+        height: 100%;
+        object-fit: contain;
+        object-position: center;
+    }
+
+}
+
+.info_product {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 100%;
+    max-width: 400px;
+    height: 500px;
+
+    a {
+        width: 100%;
+        border: 1px solid #652eac;
+        padding: 10px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        border-radius: 8px;
+        font-weight: 600;
+
+        &:hover {
+            background-color: #652eac;
+            color: white;
+        }
+    }
+
+
+}
+
+.info_product_price {
+    font-weight: 900;
+    font-size: 1.5rem;
+}
+
+.info_product_category {
+    background-color: #652eac;
+    width: 150px;
+    height: 40px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 5px;
+    font-size: 1.2rem;
+    font-weight: 600;
+}
+
+.info_product_status {
+    color: red;
+    font-weight: bold;
+}
